Package change
Version

Updated to server-core-0.1.2
