+++ build.gradle
@@ -2,8 +2,8 @@ plugins {
     id 'org.jetbrains.kotlin.jvm' version '1.3.0'
 }
 
-group 'world.gregs'
-version '0.1.0'
+group 'world.gregs.hestia'
+version '0.1.1'
 
 repositories {
     mavenCentral()
