+++ src/main/kotlin/world/gregs/hestia/ls/network/worlds/in/WorldReconnect.kt
@@ -1,12 +1,12 @@
-package world.gregs.hestia.network.worlds.`in`
+package world.gregs.hestia.ls.network.worlds.`in`
 
-import world.gregs.hestia.network.Session
-import world.gregs.hestia.network.packets.InboundPacket
-import world.gregs.hestia.network.packets.Packet
-import world.gregs.hestia.network.packets.PacketOpcode
-import world.gregs.hestia.network.packets.PacketSize
-import world.gregs.hestia.network.worlds.Worlds
-import world.gregs.hestia.network.worlds.WorldsInboundHandler
+import world.gregs.hestia.core.network.Session
+import world.gregs.hestia.core.network.packets.InboundPacket
+import world.gregs.hestia.core.network.packets.Packet
+import world.gregs.hestia.core.network.packets.PacketOpcode
+import world.gregs.hestia.core.network.packets.PacketSize
+import world.gregs.hestia.ls.network.worlds.Worlds
+import world.gregs.hestia.ls.network.worlds.WorldsInboundHandler
 
 @PacketSize(-1)
 @PacketOpcode(11)
